front: move processXmlFile() to ImportTimetableItemConfig
This small function makes things difficult to follow because
callbacks are passed back and forth. Move it with the rest of the
XML parsing code, which is already in ImportTimetableItemConfig.
This will make refactorings easier.

It's not great that the full XML parsing code sits in
ImportTimetableItemConfig, but let's fix that a tad later to ease
reviewer pain.
